pyproject.toml
[project]
name = "media-empire"
version = "0.1.0"
description = "Automated content creation system for generating and publishing YouTube videos and shorts"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    # Core video processing
    "moviepy>=1.0.3",
    # AI/ML
    "openai>=1.0.0",
    "transformers>=4.0.0",
    # Google Cloud services
    "google-auth-oauthlib>=1.0.0",
    "google-cloud-texttospeech>=2.0.0",
    "google-cloud-speech>=2.0.0",
    "google-api-python-client>=2.0.0",
    "oauth2client>=4.1.3",
    # Image/Video processing
    "Pillow>=10.0.0",
    "opencv-python>=4.0.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.11.0",
    "colorgram.py>=1.2.0",
    # Database
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    # Web framework and API
    "flask>=3.0.0",
    "flask-swagger-ui>=4.11.0",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    # Pipeline orchestration
    # Note: This project was built with ZenML 0.40.x which used BaseParameters
    # Modern ZenML removed BaseParameters - code will need updates to work
    "zenml>=0.60.0",
    "click>=8.1.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "munch>=4.0.0",
    "requests>=2.31.0",
    "httplib2>=0.22.0",
    "urllib3>=2.0.0",
    # Text-to-speech
    "gTTS>=2.4.0",
    # Social media
    "pytube>=15.0.0",
    "youtube-transcript-api>=0.6.0",
    "python-telegram-bot>=20.0",
    "psutil>=5.9.0",
    # Web scraping and parsing
    "beautifulsoup4>=4.12.0",
    "langdetect>=1.0.9",
    "pandas>=2.0.0",
    # ML/AI acceleration
    "accelerate>=0.24.0",
    "safetensors>=0.4.0",
    "python-pptx>=1.0.2",
]

[project.optional-dependencies]
stable-diffusion = [
    "diffusers>=0.24.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 150

[tool.ruff.lint]
ignore = ["E231"]

[tool.ruff.format]
quote-style = "double"

[tool.black]
line-length = 150
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 150
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true